[ie/youtube] Refactor project structure
