+++ yt_dlp/extractor/youtube/pot/_builtin/__init__.py
@@ -1,0 +1,4 @@
+from .memory_cache import MemoryLRUPCP as _MemoryLRUPCP  # noqa: F401
+from .webpo_cachespec import WebPoPCSP as _WebPoPCSP  # noqa: F401
+
+_trigger_import = None  # noqa: F401
