+++ yt_dlp/extractor/youtube/pot/_builtin/memory_cache.py
