--- yt_dlp/extractor/youtube/pot/builtin/__init__.py
@@ -1,2 +1,0 @@
-from .cache.memory import MemoryLRUPCP as _MemoryPCP  # noqa: F401
-from .cachespec.webpo import WebPoPCSP as _WebPoPCSP  # noqa: F401
